fix: dtype

+++ src/BTVNanoCommissioning/workflows/QCD_validation.py
@@ -192,8 +192,8 @@ class NanoProcessor(processor.ProcessorABC):
                 ak.values_astype(selev.luminosityBlock, np.float32),
             )
             weights.add("psweight", psweight)
-            genflavor = ak.zeros_like(sjets.pt)
-            lj_matched_JetSVs_genflav = ak.zeros_like(lj_matched_JetSVs.pt)
+            genflavor = ak.zeros_like(sjets.pt, dtype=int)
+            lj_matched_JetSVs_genflav = ak.zeros_like(lj_matched_JetSVs.pt, dtype=int)
 
         # Systematics information
         if shift_name is None:
